(fix) usdt and usdc are fixed

+++ Models/ModeOutput.cs
@@ -5,12 +5,12 @@ namespace SlippageBackend.Models
     public partial class ModelOutput
     {
         [JsonPropertyName("execution_price")]
-        public long ExecutionPrice { get; set; }
+        public decimal ExecutionPrice { get; set; }
 
         [JsonPropertyName("slippage_amount")]
-        public long SlippageAmount { get; set; }
+        public decimal SlippageAmount { get; set; }
 
         [JsonPropertyName("slippage_percentage")]
-        public long SlippagePercentage { get; set; }
+        public decimal SlippagePercentage { get; set; }
     }
 }
